+++ pczt/CHANGELOG.md
@@ -14,10 +14,11 @@ workspace.
 - `pczt::ExtractError`
 - `pczt::EffectsOnly`
 - `pczt::roles::signer`:
-  - `Signer::sighash`
   - `Signer::append_transparent_signature`
-  - `Signer::apply_sapling_signature`
   - `Signer::apply_orchard_signature`
+  - `Signer::apply_sapling_signature`
+  - `Signer::shielded_sighash`
+  - `Signer::sighash`
 
 ### Changed
 - Migrated to `orchard 0.12`, `sapling-crypto 0.6`.